+++ src/stories/tabs.stories.tsx
@@ -31,3 +31,28 @@ export const VariantUnderline = () => {
     </Stack>
   )
 }
+
+export const VariantIndicator = () => {
+  const sizes = ['md', 'lg'] as const
+  const orientations = ['horizontal'] as const
+  return (
+    <Stack spacing="12">
+      {orientations.map((orientation) => (
+        <Stack spacing="8" key={orientation}>
+          {sizes.map((size) => (
+            <Stack key={size} spacing="4">
+              <Tabs size={size} orientation={orientation} variant="indicator">
+                <TabList>
+                  <Tab>Home</Tab>
+                  <Tab>Components</Tab>
+                  <Tab>Pricing</Tab>
+                </TabList>
+                <TabIndicator />
+              </Tabs>
+            </Stack>
+          ))}
+        </Stack>
+      ))}
+    </Stack>
+  )
+}